Revert "Outside"
