+++ .eslintrc.js
@@ -17,7 +17,5 @@ module.exports = {
     'vue'
   ],
   // add your custom rules here
-  rules: {
-    'vue/no-parsing-error': [2, { 'x-invalid-end-tag': false }]
-  }
+  rules: {}
 }
