+++ assets/server/api.js
@@ -16,105 +16,87 @@ import HTTP from '~/assets/server/http.js'
 // ##############################################################################################
 // ————首页信息————
 function getshuoshuo(params) {
-  return HTTP.GET('/index', params) //获得心情
-}
-
-function deletById(params) {
-  return HTTP.GET('/modify/delete', params) //获得心情
+   return HTTP.GET('/index',params)//访问次数
 }
 // ————次数————
 function getVisitors() {
-  return HTTP.GET('/article/articleList') //访问次数
+   return HTTP.GET('/article/articleList')//访问次数
 }
 
 // ————文章————
 function getAticleList() {
-  return HTTP.GET('/article/articleList') //获得文章列表
+   return HTTP.GET('/article/articleList')//获得文章列表
 }
 
 function postAticle(params) {
-  return HTTP.POST('/article/articleContent', params) //获得文章内容
+   return HTTP.POST('/article/articleContent', params)//获得文章内容
 }
 
 function getAticleRead() {
-  return HTTP.get('/article/articleList') //获得文章阅读次数
+   return HTTP.get('/article/articleList')//获得文章阅读次数
 }
 
 // ————评论————
-function getMessageList(params) { //parames:{每页显示个数，页码}
-  return HTTP.GET('/message/messageList', params) //获得评论列表
-}
-
-function replyMessage(params) { //回复留言
-  return HTTP.GET('/modify/updateReply', params) //回复评论
+function getMessageList(params) {//parames:{每页显示个数，页码}
+   return HTTP.GET('/message/messageList', params)//获得评论列表
 }
 
 function postMessage(params) {
-  return HTTP.POST('/message/takeMessage', params) //提交评论
+   return HTTP.POST('/message/takeMessage', params)//提交评论
 }
 
 // ————统计次数————
 function addTimes(params) {
-  return HTTP.GET('/times/addTimes', params) //增加次数
+   return HTTP.GET('/times/addTimes', params)//增加次数
 }
 
 function getTimes(params) {
-  return HTTP.GET('/times/getTimes', params) //获得次数
+   return HTTP.GET('/times/getTimes', params)//获得次数
 }
 // ————项目————
 function getProjectByType(params) {
-  return HTTP.POST('/project/getProjectByType', params) //增加次数
+   return HTTP.POST('/project/getProjectByType', params)//增加次数
 }
 
 // ————上传————
 function upload(params) {
-  return HTTP.UPLOAD('/upload/img', params) //增加次数
+   return HTTP.UPLOAD('/upload/img', params)//增加次数
 }
 
 // ————发布————
 function publicShuoshuo(params) {
-  return HTTP.POST('/public/shuoshuo', params) //增加次数
+   return HTTP.POST('/public/shuoshuo', params)//增加次数
 }
 
-// ————个人中心————
-function login(params) {
-  return HTTP.POST('/person/login', params) //增加次数
-}
 
 export default {
-  // funcName
-  //    删除
-  deletById,
-  //  获得说说
-  getshuoshuo,
-  // 访问次数
-  getVisitors,
+   // funcName
+   //  获得说说
+   getshuoshuo,
+   // 访问次数
+   getVisitors,
 
-  // 博客接口
-  getAticleList,
-  postAticle,
-  getAticleRead,
+   // 博客接口
+   getAticleList,
+   postAticle,
+   getAticleRead,
 
-  // 留言接口
-  getMessageList,
-  postMessage,
-  replyMessage,
+   // 留言接口
+   getMessageList,
+   postMessage,
 
-  //  统计次数
-  addTimes,
-  getTimes,
+   //  统计次数
+   addTimes,
+   getTimes,
 
-  // 项目
-  getProjectByType,
+   // 项目
+   getProjectByType,
 
-  // 上传
-  upload,
+   // 上传
+   upload,
 
-  // 发表
-  publicShuoshuo,
-  
-  // 个人中心
-  login,
+   // 发表
+   publicShuoshuo,
 }
 
 // ##############################################################################################
